requirements.txt
requests>=2.31.0
rich>=13.7.0
click>=8.1.7
python-dotenv>=1.0.0
tabulate>=0.9.0
prompt-toolkit>=3.0.43